refactor: refactor Core files
