+++ dristi/lib/src/core/base/base_stateful_widget.dart
@@ -1,5 +1,5 @@
 import 'package:dristi/src/core/styles/colors/colors.dart';
-import 'package:dristi/src/core/styles/texts.dart';
+import 'package:dristi/src/core/styles/texts/texts.dart';
 import 'package:flutter/material.dart';
 
 abstract class BaseStatefulWidget<T extends StatefulWidget> extends State<T> {
