+++ dristi/lib/src/core/routes/app_router.dart
@@ -3,6 +3,7 @@ import 'package:dristi/src/core/global_widgets/web_view_screen.dart';
 import 'package:dristi/src/core/routes/app_routes.dart';
 import 'package:dristi/src/core/routes/navigation_helper.dart';
 import 'package:dristi/src/core/utils/localization_ext.dart';
+import 'package:dristi/src/features/accommodations_list/presentation/pages/accommodations_list_screen.dart';
 import 'package:dristi/src/features/destination/domain/entities/destination_entities.dart';
 import 'package:dristi/src/features/destinations_list/presentation/pages/destinations_list_screen.dart';
 import 'package:dristi/src/features/districts/presentation/pages/districts_list_screen.dart';
@@ -11,7 +12,6 @@ import 'package:dristi/src/features/gallery/presentation/pages/image_view_screen
 import 'package:dristi/src/features/home/advertisements/domain/entity/advertisement_entity.dart';
 import 'package:dristi/src/features/home/home_screen/pages/home_screen.dart';
 import 'package:dristi/src/features/accommodation/presentation/pages/accommodation_screen.dart';
-import 'package:dristi/src/features/accommodations_list/presentation/pages/accommodations_list_screen.dart';
 import 'package:dristi/src/features/on_boarding/presentation/pages/on_boarding_screen.dart';
 import 'package:dristi/src/features/settings/presentation/pages/settings_screen.dart';
 import 'package:dristi/src/features/destination/presentation/pages/destination_screen.dart';
