+++ dristi/lib/src/features/destinations_list/presentation/widgets/destinations_list.dart
@@ -34,7 +34,7 @@ class _DestinationsListState
 
     if (destinationModelsItems.status != DestinationListStatus.success ||
         destinationModelsItems.data == null) {
-      return buildDestinationListShimmer(context);
+      return buildDestinationsListShimmer(context);
     }
 
     List<DestinationsListEntity> fetchResult = searchDestinations();
