+++ dristi/lib/src/features/home/popular_districts/presentations/widgets/popular_districts_builder.dart
@@ -58,7 +58,7 @@ class _PopularCitiesBuilderState
 
     if (popularDistricts.status != PopularDistrictsStatus.success ||
         popularDistricts.data == null) {
-      return buildPopularDistrictsShimmer(context);
+      return buildHomeScreenPopularDistrictsShimmer(context);
     }
 
     return SizedBox(
