+++ dristi/lib/src/features/home/top_destinations/presentations/widgets/top_destinations_builder.dart
@@ -59,7 +59,7 @@ class _TopDestinationBuilderState
 
     if (topDestinationItems.status != TopDestinationsStatus.success ||
         topDestinationItems.data == null) {
-      return buildTopDestinationsShimmer(context);
+      return buildHomeScreenTopDestinationsShimmer(context);
     }
 
     return SizedBox(
